+++ code/src/Plexdata.LogWriter.Console.Tests/Settings/ConsoleLoggerSettingsTests.cs
@@ -22,6 +22,8 @@
  * SOFTWARE.
  */
 
+using Microsoft.Extensions.Configuration;
+using Moq;
 using NUnit.Framework;
 using Plexdata.LogWriter.Definitions;
 using Plexdata.LogWriter.Settings;
@@ -35,6 +37,28 @@ namespace Plexdata.LogWriter.Console.Tests.Settings
     [TestOf(nameof(ConsoleLoggerSettings))]
     public class ConsoleLoggerSettingsTests
     {
+        #region Prologue
+
+        private Mock<IConfigurationSection> section;
+        private Mock<IConfiguration> configuration;
+
+        [SetUp]
+        public void Setup()
+        {
+            this.section = new Mock<IConfigurationSection>();
+            this.configuration = new Mock<IConfiguration>();
+
+            this.configuration
+                .Setup(x => x.GetSection(It.IsAny<String>()))
+                .Returns(this.section.Object);
+
+            this.section
+                .Setup(x => x.GetSection(It.IsAny<String>()))
+                .Returns(this.section.Object);
+        }
+
+        #endregion
+
         #region ConsoleLoggerSettings
 
         [Test]
@@ -65,6 +89,201 @@ namespace Plexdata.LogWriter.Console.Tests.Settings
             Assert.That(instance.Coloring[LogLevel.Fatal].Background, Is.EqualTo(ConsoleColor.DarkRed));
             Assert.That(instance.Coloring[LogLevel.Critical].Foreground, Is.EqualTo(ConsoleColor.Black));
             Assert.That(instance.Coloring[LogLevel.Critical].Background, Is.EqualTo(ConsoleColor.Red));
+        }
+
+        [Test]
+        public void ConsoleLoggerSettings_ValidateDefaultSettingsForInvalidConfiguration_DefaultSettingsAsExpected()
+        {
+            ConsoleLoggerSettings instance = new ConsoleLoggerSettings(null);
+
+            Assert.That(instance.UseColors, Is.True);
+            Assert.That(instance.WindowTitle, Is.Empty);
+            Assert.That(instance.QuickEdit, Is.False);
+            Assert.That(instance.BufferSize.IsValid, Is.False);
+            Assert.That(instance.BufferSize.Width, Is.Zero);
+            Assert.That(instance.BufferSize.Lines, Is.Zero);
+            Assert.That(instance.Coloring.Count, Is.EqualTo(8));
+            Assert.That(instance.Coloring[LogLevel.Trace].Foreground, Is.EqualTo(ConsoleColor.Gray));
+            Assert.That(instance.Coloring[LogLevel.Trace].Background, Is.EqualTo(ConsoleColor.Black));
+            Assert.That(instance.Coloring[LogLevel.Debug].Foreground, Is.EqualTo(ConsoleColor.Gray));
+            Assert.That(instance.Coloring[LogLevel.Debug].Background, Is.EqualTo(ConsoleColor.Black));
+            Assert.That(instance.Coloring[LogLevel.Verbose].Foreground, Is.EqualTo(ConsoleColor.White));
+            Assert.That(instance.Coloring[LogLevel.Verbose].Background, Is.EqualTo(ConsoleColor.Black));
+            Assert.That(instance.Coloring[LogLevel.Message].Foreground, Is.EqualTo(ConsoleColor.White));
+            Assert.That(instance.Coloring[LogLevel.Message].Background, Is.EqualTo(ConsoleColor.Black));
+            Assert.That(instance.Coloring[LogLevel.Warning].Foreground, Is.EqualTo(ConsoleColor.Yellow));
+            Assert.That(instance.Coloring[LogLevel.Warning].Background, Is.EqualTo(ConsoleColor.Black));
+            Assert.That(instance.Coloring[LogLevel.Error].Foreground, Is.EqualTo(ConsoleColor.Red));
+            Assert.That(instance.Coloring[LogLevel.Error].Background, Is.EqualTo(ConsoleColor.Black));
+            Assert.That(instance.Coloring[LogLevel.Fatal].Foreground, Is.EqualTo(ConsoleColor.Gray));
+            Assert.That(instance.Coloring[LogLevel.Fatal].Background, Is.EqualTo(ConsoleColor.DarkRed));
+            Assert.That(instance.Coloring[LogLevel.Critical].Foreground, Is.EqualTo(ConsoleColor.Black));
+            Assert.That(instance.Coloring[LogLevel.Critical].Background, Is.EqualTo(ConsoleColor.Red));
+        }
+
+        [Test]
+        public void ConsoleLoggerSettings_ConfigurationValid_GetSectionCalledAsExpected()
+        {
+            ConsoleLoggerSettings instance = new ConsoleLoggerSettings(this.configuration.Object);
+
+            this.configuration.Verify(x => x.GetSection("Plexdata:LogWriter:Settings"), Times.Exactly(2));
+            this.section.Verify(x => x.GetSection("BufferSize"), Times.Once);
+            this.section.Verify(x => x.GetSection("Coloring"), Times.Once);
+        }
+
+        [Test]
+        [TestCase(null, true)]
+        [TestCase("", true)]
+        [TestCase(" ", true)]
+        [TestCase("invalid", true)]
+        [TestCase("true", true)]
+        [TestCase("TRUE", true)]
+        [TestCase("True", true)]
+        [TestCase("false", false)]
+        [TestCase("FALSE", false)]
+        [TestCase("False", false)]
+        public void ConsoleLoggerSettings_ConfigurationValid_GetSectionValueForUseColorsAsExpected(String value, Object expected)
+        {
+            this.section.SetupGet(x => x["UseColors"]).Returns(value);
+
+            ConsoleLoggerSettings instance = new ConsoleLoggerSettings(this.configuration.Object);
+
+            Assert.That(instance.UseColors, Is.EqualTo(expected));
+        }
+
+        [Test]
+        [TestCase(null, "")]
+        [TestCase("", "")]
+        [TestCase(" ", "")]
+        [TestCase("WindowTitle", "WindowTitle")]
+        [TestCase("Window Title", "Window Title")]
+        [TestCase("WINDOWTITLE", "WINDOWTITLE")]
+        [TestCase("WINDOW TITLE", "WINDOW TITLE")]
+        public void ConsoleLoggerSettings_ConfigurationValid_GetSectionValueForWindowTitleAsExpected(String value, Object expected)
+        {
+            this.section.SetupGet(x => x["WindowTitle"]).Returns(value);
+
+            ConsoleLoggerSettings instance = new ConsoleLoggerSettings(this.configuration.Object);
+
+            Assert.That(instance.WindowTitle, Is.EqualTo(expected));
+        }
+
+        [Test]
+        [TestCase(null, false)]
+        [TestCase("", false)]
+        [TestCase(" ", false)]
+        [TestCase("invalid", false)]
+        [TestCase("true", true)]
+        [TestCase("TRUE", true)]
+        [TestCase("True", true)]
+        [TestCase("false", false)]
+        [TestCase("FALSE", false)]
+        [TestCase("False", false)]
+        public void ConsoleLoggerSettings_ConfigurationValid_GetSectionValueForQuickEditAsExpected(String value, Object expected)
+        {
+            this.section.SetupGet(x => x["QuickEdit"]).Returns(value);
+
+            ConsoleLoggerSettings instance = new ConsoleLoggerSettings(this.configuration.Object);
+
+            Assert.That(instance.QuickEdit, Is.EqualTo(expected));
+        }
+
+        [Test]
+        [TestCase(null, null, 0, 0)]
+        [TestCase("", "", 0, 0)]
+        [TestCase(" ", " ", 0, 0)]
+        [TestCase("invalid", "invalid", 0, 0)]
+        [TestCase(null, "23", 0, 0)]
+        [TestCase("", "23", 0, 0)]
+        [TestCase(" ", "23", 0, 0)]
+        [TestCase("invalid", "23", 0, 0)]
+        [TestCase("42", null, 0, 0)]
+        [TestCase("42", "", 0, 0)]
+        [TestCase("42", " ", 0, 0)]
+        [TestCase("42", "invalid", 0, 0)]
+        [TestCase("0", "0", 0, 0)]
+        [TestCase("42", "23", 42, 23)]
+        [TestCase("42.0", "23.5", 0, 0)]
+        public void ConsoleLoggerSettings_ConfigurationValid_GetSectionValueForBufferSizeAsExpected(String sWidth, String sLines, Int32 nWidth, Int32 nLines)
+        {
+            Mock<IConfigurationSection> dimension = new Mock<IConfigurationSection>();
+
+            this.section.Setup(x => x.GetSection("BufferSize")).Returns(dimension.Object);
+
+            dimension.SetupGet(x => x["Width"]).Returns(sWidth);
+            dimension.SetupGet(x => x["Lines"]).Returns(sLines);
+
+            ConsoleLoggerSettings instance = new ConsoleLoggerSettings(this.configuration.Object);
+
+            Assert.That(instance.BufferSize.Width, Is.EqualTo(nWidth));
+            Assert.That(instance.BufferSize.Lines, Is.EqualTo(nLines));
+        }
+
+        [Test]
+        [TestCase(LogLevel.Trace, null, null, ConsoleColor.Gray, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "", "", ConsoleColor.Gray, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, " ", " ", ConsoleColor.Gray, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "red", null, ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "red", "", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "red", " ", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "RED", null, ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "RED", "", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "RED", " ", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "Red", null, ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "Red", "", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, "Red", " ", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Trace, null, "green", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, "", "green", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, " ", "green", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, null, "GREEN", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, "", "GREEN", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, " ", "GREEN", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, null, "Green", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, "", "Green", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, " ", "Green", ConsoleColor.Gray, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, "red", "green", ConsoleColor.Red, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, "RED", "GREEN", ConsoleColor.Red, ConsoleColor.Green)]
+        [TestCase(LogLevel.Trace, "Red", "Green", ConsoleColor.Red, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, null, null, ConsoleColor.Yellow, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "", "", ConsoleColor.Yellow, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, " ", " ", ConsoleColor.Yellow, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "red", null, ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "red", "", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "red", " ", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "RED", null, ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "RED", "", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "RED", " ", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "Red", null, ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "Red", "", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, "Red", " ", ConsoleColor.Red, ConsoleColor.Black)]
+        [TestCase(LogLevel.Warning, null, "green", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, "", "green", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, " ", "green", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, null, "GREEN", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, "", "GREEN", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, " ", "GREEN", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, null, "Green", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, "", "Green", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, " ", "Green", ConsoleColor.Yellow, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, "red", "green", ConsoleColor.Red, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, "RED", "GREEN", ConsoleColor.Red, ConsoleColor.Green)]
+        [TestCase(LogLevel.Warning, "Red", "Green", ConsoleColor.Red, ConsoleColor.Green)]
+        public void ConsoleLoggerSettings_ConfigurationValid_GetSectionValueForColoringAsExpected(LogLevel level, String sForeground, String sBackground, ConsoleColor nForeground, ConsoleColor nBackground)
+        {
+            Mock<IConfigurationSection> parent = new Mock<IConfigurationSection>();
+            Mock<IConfigurationSection> coloring = new Mock<IConfigurationSection>();
+            Mock<IConfigurationSection> others = new Mock<IConfigurationSection>();
+
+            this.section.Setup(x => x.GetSection("Coloring")).Returns(parent.Object);
+            parent.Setup(x => x.GetSection(It.IsAny<String>())).Returns((String current) => { return (current == level.ToString()) ? coloring.Object : others.Object; });
+
+            coloring.SetupGet(x => x["Foreground"]).Returns(sForeground);
+            coloring.SetupGet(x => x["Background"]).Returns(sBackground);
+
+            ConsoleLoggerSettings instance = new ConsoleLoggerSettings(this.configuration.Object);
+
+            Assert.That(instance.Coloring[level].Foreground, Is.EqualTo(nForeground));
+            Assert.That(instance.Coloring[level].Background, Is.EqualTo(nBackground));
         }
 
         #endregion